move ollama tools to tools.py
